22.09.21.03 - task, ans 파일 시퀀스 적용
