+++ src/soobin/answer/chap01/exam01/Ans01.java
@@ -1,6 +1,6 @@
 package soobin.answer.chap01.exam01;
 
-public class ans01 {
+public class Ans01 {
 
 	public static void main(String[] args) {
 		//1. 콘솔창에 Hello World를 출력해보세요.
