+++ src/soobin/answer/chap02/exam01/Ans02.java
@@ -1,16 +1,16 @@
-package soobin.assign.chap02.exam01;
+package soobin.answer.chap02.exam01;
 
-public class Task01 {
+public class Ans02 {
 
 	public static void main(String[] args) {
 		
 		
 		//1. 아래 구문의 실행 결과로 3이 출력되도록 해보세요.
 		//코드 작성 부분
-		
+		int num1 = 3;
 		
 		//주석해제하여 실행
-//		System.out.println(num1);
+		System.out.println(num1);
 		
 		
 		
@@ -23,13 +23,13 @@ public class Task01 {
 		// 수학 | 55
 		//----------
 		
-		int korean;
-		int english;
-		int math;
+		int korean = 80;
+		int english = 75;
+		int math = 55;
 		
-		int avg;
+		int avg = (korean + english + math) / 3;
 		
-//		System.out.println(avg);
+		System.out.println(avg);
 		
 			
 		
@@ -37,6 +37,8 @@ public class Task01 {
 		//   완료했다면 다른 자연수를 넣어도 알맞게 나오는지 확인해보자.
 		
 		int num2 = 10;
+		
+		System.out.println((num2) % 2); //짝수이면 나머지가 0
 		
 		
 		
@@ -49,6 +51,8 @@ public class Task01 {
 		int tmp = x;
 		
 		//(1)
+		x = y;
+		y = tmp;
 		
 		System.out.println("x=" + x); // x=2
 		System.out.println("y=" + y); // y=1		
@@ -60,7 +64,7 @@ public class Task01 {
 		int numPlus = 0;
 		
 		for(int i = 0; i < 10; i++) {
-//			System.out.print(numPlus);
+			System.out.print(numPlus++);
 		}
 		
 		
@@ -71,7 +75,7 @@ public class Task01 {
 		int numMinus = 10;
 		
 		for(int i = 0; i < 10; i++) {
-//			System.out.print(numMinus);
+			System.out.print(--numMinus);
 		}
 		
 		
@@ -83,10 +87,11 @@ public class Task01 {
 		
 		
 		//출력예제
+		boolean printBool = true;
 		
-//		if(printBool) {
-//			System.out.println(printStr);
-//		}
+		if(printBool) {
+			System.out.println(printStr);
+		}
 		
 		
 		
@@ -96,6 +101,9 @@ public class Task01 {
 		
 		String pin1 = "881120-1068234";
 		
+		System.out.println(pin1.substring(0, 6));
+		System.out.println(pin1.substring(7));
+		
 		
 				
 		// 9. 주민등록번호 뒷자리의 맨 첫 번째 숫자는 성별을 나타낸다. 주민등록번호에서 성별을 나타내는 숫자를 출력해 보자.
@@ -103,11 +111,15 @@ public class Task01 {
 		
 		String pin2 = "881120-1068234";
 		
+		System.out.println(pin2.charAt(7));
+		
 		
 		
 		// 10. 다음과 같은 문자열 a:b:c:d가 있다. 문자열의 replace 함수를 사용하여 a#b#c#d로 바꿔서 출력해 보자.
 		
 		String abcd = "a:b:c:d";
+		
+		System.out.println(abcd.replace(":", "#"));
